+++ src/__tests__/App.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { render } from "utils/test-utils";
+import { render } from "utils/utils.helper";
 import App from "../App";
 
 it("should render without crashing", () => {