+++ src/components/Footer/components/Nav.tsx
@@ -1,27 +1,24 @@
-import React from "react";
+import React, { memo } from "react";
 import SLink from "components/SLink";
 import { styled } from "@mui/material";
 
-const Nav: React.FC = () => {
-  return (
-    <StyledNav>
-      <SLink external href="https://twitter.com/">
-        Twitter
-      </SLink>
-      <SLink external href="https://discord.gg">
-        Discord
-      </SLink>
-      <SLink external href="https://medium.com">
-        Medium
-      </SLink>
-      <SLink external href="https://github.com/">
-        ToS
-      </SLink>
-    </StyledNav>
-  );
-};
-
-const StyledNav = styled("nav")(({ theme }) => ({
+const Nav: React.FC = () => (
+  <StyledNav>
+    <SLink external href="https://twitter.com/">
+      Twitter
+    </SLink>
+    <SLink external href="https://discord.gg">
+      Discord
+    </SLink>
+    <SLink external href="https://medium.com">
+      Medium
+    </SLink>
+    <SLink external href="https://github.com/">
+      ToS
+    </SLink>
+  </StyledNav>
+);
+const StyledNav = styled("nav")(() => ({
   alignItems: "center",
   display: "flex",
   flexWrap: "wrap",
@@ -31,4 +28,4 @@ const StyledNav = styled("nav")(({ theme }) => ({
   },
 }));
 
-export default React.memo(Nav);
+export default memo(Nav);