+++ src/contexts/AccountContext/Provider.tsx
@@ -1,12 +1,12 @@
-import React, { useCallback, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import Context from "./Context";
 import { generateNewWallet } from "utils/wallet";
 
 const AccountProvider: React.FC = ({ children }) => {
   const [accountRs, setAccountRs] = useState<string>();
   const [accountSeed, setAccountSeed] = useState<string>();
-
-  // creates a new seed, converts to accountRs format, sets it in state
+  const [accountAlias, setAccountAlias] = useState<string>();
+  // Creates a new seed, converts to accountRs format, sets it in state
   const fetchNewAccount = useCallback(async () => {
     const { accountRs, accountSeed } = await generateNewWallet();
 
@@ -14,10 +14,17 @@ const AccountProvider: React.FC = ({ children }) => {
 
     setAccountSeed(accountSeed);
   }, []);
-
-  // flushes seed back to empty string after we're done using it
+  const handleLogin = useCallback((account: string) => {
+    setAccountRs(account);
+  }, []);
+  // Flushes seed back to empty string after we're done using it
   const flushAccountSeed = useCallback(() => {
     setAccountSeed("");
+  }, []);
+
+  useEffect(() => {
+    // TODO: fetch the actual alias using a utils call
+    setAccountAlias("testAlias123");
   }, []);
 
   return (
@@ -25,8 +32,10 @@ const AccountProvider: React.FC = ({ children }) => {
       value={{
         accountRs,
         accountSeed,
+        accountAlias,
         fetchFn: fetchNewAccount,
         flushFn: flushAccountSeed,
+        userLogin: handleLogin,
       }}
     >
       {children}