+++ src/global.d.ts
@@ -1,0 +1,5 @@
+/*
+ * Declare global values here, which are available throughout the codebase.
+ * these get set via Vitejs, within the vite.config.ts file
+ */
+declare const APP_VERSION: string;