+++ src/utils/__tests__/generatePassphrase.test.js
@@ -10,8 +10,7 @@ let getRandomFn;
 function mockWindowCrypto(randomness) {
   getRandomFn = jest.fn((random) => {
     random.forEach((el, index, array) => {
-      array[index] =
-        typeof randomness === "number" ? randomness : randomness[index]; // anything other than 0 should produce a different passphrase
+      array[index] = typeof randomness === "number" ? randomness : randomness[index]; // anything other than 0 should produce a different passphrase
     });
   });
 
@@ -69,9 +68,7 @@ it("should generate a predictable passphrase based on specific randomness", () =
   const passphrase = generatePassPhrase();
 
   // NEVER USE THIS SEED
-  expect(passphrase).toEqual(
-    "beauty fail everytime glove corner content realize witch secret grace rose finally"
-  );
+  expect(passphrase).toBe("beauty fail everytime glove corner content realize witch secret grace rose finally");
   expect(getRandomFn).toHaveBeenCalledTimes(2);
 });
 
@@ -87,9 +84,7 @@ it("should error when crypto isn't in the window", () => {
 it("should have the same wordlist as before", () => {
   mockWindowCrypto(0);
   const wordHash = createHash("sha256").update(wordList.join("")).digest("hex");
-  expect(wordHash).toEqual(
-    "36457fdcaaff463a84891cf5a32710cd754b7efca67060163dd5ae1e5ffe2934"
-  );
+  expect(wordHash).toBe("36457fdcaaff463a84891cf5a32710cd754b7efca67060163dd5ae1e5ffe2934");
 });
 
 it("should call crypto only twice", () => {