+++ src/views/Login/components/OnboardingStepper/components/3.ReEnterSeedStep/ReEnterSeedStep.tsx
@@ -1,19 +1,18 @@
-import React, { ReactElement, useCallback, useEffect, useMemo, useState } from "react";
+import React, { ReactElement, memo, useCallback, useEffect, useMemo, useState } from "react";
 import useAccount from "hooks/useAccount";
-import { Alert, Chip, Grid, styled, Typography, Button } from "@mui/material";
+import { Alert, Button, Chip, Grid, Typography, styled } from "@mui/material";
 import { IStepProps } from "../types";
 import useBreakpoint from "hooks/useBreakpoint";
 import RemoveCircleOutlineIcon from "@mui/icons-material/RemoveCircleOutline";
 
 interface IReEntryChipProps {
-  onClickFn: Function;
-  onDeleteFn: Function;
+  onClickFn: (label: string) => boolean;
+  onDeleteFn: (label: string) => boolean;
   labelText: string;
 }
 
 const ReEntryChip: React.FC<IReEntryChipProps> = ({ onClickFn, onDeleteFn, labelText }) => {
   const [isClicked, setIsClicked] = useState(false);
-
   const handleDelete = useCallback(() => {
     const result = onDeleteFn(labelText);
 
@@ -21,7 +20,6 @@ const ReEntryChip: React.FC<IReEntryChipProps> = ({ onClickFn, onDeleteFn, label
       setIsClicked((prev) => !prev);
     }
   }, [onDeleteFn, labelText]);
-
   const handleClick = useCallback(() => {
     const result = onClickFn(labelText);
     if (result) {
@@ -37,7 +35,6 @@ const ReEntryChip: React.FC<IReEntryChipProps> = ({ onClickFn, onDeleteFn, label
     <StyledChip onClick={handleClick} label={labelText} />
   );
 };
-
 const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
   const { accountSeed } = useAccount();
   const [reEntryText, setReEntryText] = useState<Array<string>>();
@@ -45,16 +42,17 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
   const [wordList, setWordList] = useState<Array<string>>();
   const isSmallBrowser = useBreakpoint("<", "sm");
   const [reEntryCounter, setReEntryCounter] = useState<number | undefined>(0);
+  /*
+   * DEV USE
+   * Skips past the re-entry step for convenience
+   * UseEffect(() => {
+   *   StepForwardFn(); // steps past this step altogether without having to re-enter words
+   * }, [stepForwardFn]);
+   */
 
-  // DEV USE
-  // skips past the re-entry step for convenience
-  // useEffect(() => {
-  //   stepForwardFn(); // steps past this step altogether without having to re-enter words
-  // }, [stepForwardFn]);
-
-  // removes from reEntryText when a user clicks the "X" on chips
+  // Removes from reEntryText when a user clicks the "X" on chips
   const reEntryDeleteFn = useCallback((label: string) => {
-    let didChange: boolean = false;
+    let didChange = false;
     setReEntryText((prev) => {
       let newArray: Array<string> = [];
 
@@ -63,31 +61,33 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
       }
 
       const len = prev?.length || 0;
-      // TODO: mark the last word picked somewhere on the chip (asterisk?) so the user can more easily find the last word to remove
-      // they've clicked the last label in the array, this is the only condition we want to allow
-      // the user to remove the word
+
+      /*
+       * TODO: mark the last word picked somewhere on the chip (asterisk?) so the user can more easily find the last word to remove
+       * They've clicked the last label in the array, this is the only condition we want to allow
+       * The user to remove the word
+       */
       if (prev[len - 1] === label) {
         newArray = [...prev.slice(0, len - 1)];
         didChange = true;
         return newArray;
       }
 
-      // otherwise we give them the same array back
+      // Otherwise we give them the same array back
       return prev;
     });
     return didChange;
   }, []);
-
-  // adds to reEntryText when a user clicks chips
+  // Adds to reEntryText when a user clicks chips
   const reEntryChipClickFn = useCallback((label: string) => {
-    let didChange: boolean = false;
+    let didChange = false;
     setReEntryText((prev) => {
       let newArray: Array<string> = [];
       if (prev === undefined) {
         prev = [];
       }
 
-      // the user has already re-entered this word
+      // The user has already re-entered this word
       if (prev.includes(label)) {
         return prev;
       }
@@ -96,16 +96,15 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
       return newArray;
     });
 
-    // the reEntryText value changed
+    // The reEntryText value changed
     return didChange;
   }, []);
-
   const AlertReEntryStatus = useMemo(() => {
     if (reEntryCounter === undefined) {
       return <></>;
     }
 
-    // user has reentered all words but they're not matching the original
+    // User has reentered all words but they're not matching the original
     if (isReEnteredCorrectly === false && reEntryCounter === 12) {
       return (
         // TODO: conditionally render the warning (don't display as the user is re-entering until they have entered all 12 words)
@@ -116,7 +115,7 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
       );
     }
 
-    // they aren't done entering all of the words yet
+    // They aren't done entering all of the words yet
     if (reEntryCounter < 12) {
       return (
         <>
@@ -141,25 +140,24 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
       </>
     );
   }, [isReEnteredCorrectly, stepForwardFn, reEntryCounter, reEntryText]);
-
-  // create chips for each word in the accountSeed and shuffle them before displaying
-  // change so seedToWordArray() is called by a useEffect with accountSeed as its dep
-  // put shuffled word list into state
+  /*
+   * Create chips for each word in the accountSeed and shuffle them before displaying
+   * Change so seedToWordArray() is called by a useEffect with accountSeed as its dep
+   * Put shuffled word list into state
+   */
   const WordChips: Array<ReactElement> | React.ReactFragment = useMemo(() => {
     if (wordList === undefined) {
       return <></>;
     }
 
-    return wordList?.map((item, index) => {
-      return (
-        <Grid item xs={4} sm={3} key={item + index}>
-          <ReEntryChip labelText={item} onClickFn={reEntryChipClickFn} onDeleteFn={reEntryDeleteFn} />
-        </Grid>
-      );
-    });
+    return wordList?.map((item, index) => (
+      <Grid item xs={4} sm={3} key={item + index}>
+        <ReEntryChip labelText={item} onClickFn={reEntryChipClickFn} onDeleteFn={reEntryDeleteFn} />
+      </Grid>
+    ));
   }, [reEntryChipClickFn, wordList, reEntryDeleteFn]);
 
-  // whenever reEntryText changes, reruns to see if matching condition is met yet and updates current counter
+  // Whenever reEntryText changes, reruns to see if matching condition is met yet and updates current counter
   useEffect(() => {
     const reEntryLength = reEntryText?.length;
 
@@ -174,8 +172,8 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
       return;
     }
     const wl = seedToWordArray(accountSeed);
-    shuffleWordlist(wl); // shuffles the wordlist in-line
-    // set shuffled word list state here from wordlist (post-shuffle)
+    shuffleWordlist(wl); // Shuffles the wordlist in-line
+    // Set shuffled word list state here from wordlist (post-shuffle)
     setWordList(wl);
   }, [accountSeed]);
 
@@ -200,40 +198,39 @@ const ReEnterSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
     </>
   );
 };
-
 const StyledButton = styled(Button)(({ theme }) => ({
   margin: theme.spacing(1),
   borderRadius: theme.shape.borderRadius,
   width: "40%",
 }));
-
 const StyledChip = styled(Chip)(({ theme }) => ({
   borderRadius: theme.shape.borderRadius,
   width: "100px",
   margin: "10px",
 }));
-
 const StyledGridContainer = styled(Grid)(({ theme }) => ({
   color: theme.palette.primary.contrastText,
   margin: "auto",
   maxWidth: "500px",
 }));
 
-//
-// helper functions
-//
+/*
+ *
+ * Helper functions
+ *
+ */
 
 function seedToWordArray(seed: string) {
   return seed.split(" ") as Array<string>;
 }
 
 function shuffleWordlist(array: Array<string>) {
-  for (var i = array.length - 1; i > 0; i--) {
-    var j = Math.floor(Math.random() * (i + 1));
-    var temp = array[i];
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    const temp = array[i];
     array[i] = array[j];
     array[j] = temp;
   }
 }
 
-export default React.memo(ReEnterSeedStep);
+export default memo(ReEnterSeedStep);