+++ src/views/Login/components/OnboardingStepper/components/types.ts
@@ -1,3 +1,3 @@
 export interface IStepProps {
-  stepForwardFn: Function;
+  stepForwardFn: () => void;
 }