+++ src/views/Login/components/RememberMeCheckbox/RememberMeCheckbox.tsx
@@ -1,17 +1,15 @@
-import { Checkbox, CheckboxProps } from "@mui/material";
-import React, { useCallback, useState } from "react";
+import { Checkbox } from "@mui/material";
+import { memo, useCallback, useState } from "react";
 
-interface IRememberMeCheckboxProps extends CheckboxProps {
-  fetchIsRememberedFn: Function;
+interface IRememberMeCheckboxProps {
+  fetchIsRememberedFn: (checked: boolean) => void;
 }
-
-const RememberMeCheckbox: React.FC<IRememberMeCheckboxProps> = ({ fetchIsRememberedFn }) => {
+const RememberMeCheckbox = ({ fetchIsRememberedFn }: IRememberMeCheckboxProps): JSX.Element => {
   const [isRemembered, setIsRemembered] = useState<boolean>(false);
-
   const handleRememberAccount = useCallback(
     (event: React.ChangeEvent<HTMLInputElement>) => {
-      fetchIsRememberedFn(!!event.target.checked); // !! cast to bool
-      setIsRemembered(!!event.target.checked);
+      fetchIsRememberedFn(Boolean(event.target.checked));
+      setIsRemembered(Boolean(event.target.checked));
     },
     [setIsRemembered, fetchIsRememberedFn]
   );
@@ -19,4 +17,4 @@ const RememberMeCheckbox: React.FC<IRememberMeCheckboxProps> = ({ fetchIsRemembe
   return <Checkbox checked={isRemembered} onChange={handleRememberAccount} />;
 };
 
-export default React.memo(RememberMeCheckbox);
+export default memo(RememberMeCheckbox);